Ajustes no tipo

+++ include/continental/hydrotools/domain/HeuristicSinkRemovalCommandInput.h
@@ -42,7 +42,7 @@ public:
     float getWeightFunctionG() const;
 
     /// Função Getter do atributo tipo de algoritmo de processamento
-    HeuristicSinkRemoval<short>::ProcessingMode getProcessingAlgorithm() const;
+    HeuristicSinkRemoval<float>::ProcessingMode getProcessingAlgorithm() const;
 
     /// Função Getter do arquivo de entrada da Superfície inicial.
     FileCommand getDemInput() const;
@@ -71,7 +71,7 @@ protected:
 
     /// Função para atualizar o valor do atributo weightFunctionG
     /// @param weightFunctionG com valor do tipo size_t
-    void setProcessingAlgorithm (const HeuristicSinkRemoval<short>::ProcessingMode processingAlgorithm);
+    void setProcessingAlgorithm (const HeuristicSinkRemoval<float>::ProcessingMode processingAlgorithm);
 
     /// Função para atualizar o valor do arquivo de entrada da Superfície inicial.
     /// @param FileCommand com valor do arquivo e nome do dataset do hdf5
@@ -96,7 +96,7 @@ private:
     float m_weightFunctionG;
 
     /// Atributo
-    HeuristicSinkRemoval<short>::ProcessingMode m_processingAlgorithm;
+    HeuristicSinkRemoval<float>::ProcessingMode m_processingAlgorithm;
 
     /// Atributo
     FileCommand m_demInput;
